Imporved error messages and refactorings

+++ tipps/api/v1.py
@@ -23,22 +23,28 @@ v1 = Blueprint('api/v1', __name__, url_prefix='/api/v1')
 @v1.route('/tokens/<string:token>/tipps')
 def list_tipps(token=None):
 	details = request.args.get('details', default=False, type=bool)
-	
+
 	sort = request.args.get('sort', default='created_desc', type=str).lower().split('_')
-	if sort[0] not in ['created','compiled','id']:
-		sort[0] = 'created'
-	if sort[1] not in ['asc', 'desc']:
+	if sort[0] == 'rand':
+		sort[0] = 'RANDOM()'
+	elif sort[0] not in ['created','compiled','id']:
+		sort[0] = 'tipp.created'
+	else:
+		sort[0] = f'tipp.{sort[0]}'
+	if len(sort) < 2:
+		sort.append('asc')
+	elif sort[1] not in ['asc', 'desc']:
 		sort[1] = 'asc'
-	sort = f'tipp.{sort[0]} {sort[1].upper()}'
-	
+	sort = f'{sort[0]} {sort[1].upper()}'
+
 	limit = request.args.get('limit', default=20, type=int)
 	offset = request.args.get('offset', default=0, type=int)
 
 	db = get_db()
 	if token is None:
-		result = db.execute(f'SELECT id,created,template FROM tipp LIMIT {limit} OFFSET {offset} ORDER BY {sort}').fetchall()
+		result = db.execute(f'SELECT id,created,template FROM tipp ORDER BY {sort} LIMIT {limit} OFFSET {offset}').fetchall()
 	else:
-		result = db.execute('SELECT tipp.id,tipp.created,tipp.template FROM tipp INNER JOIN user ON user.id = tipp.user_id WHERE user.token = ? LIMIT {limit} OFFSET {offset} ORDER BY {sort}', (token,)).fetchall()
+		result = db.execute(f'SELECT tipp.id,tipp.created,tipp.template FROM tipp INNER JOIN user ON user.id = tipp.user_id WHERE user.token = ? ORDER BY {sort} LIMIT {limit} OFFSET {offset}', (token,)).fetchall()
 
 	tipps = []
 	if not details:
@@ -64,9 +70,11 @@ def tipp_details(id):
 		tipp = {
 			'id': id,
 			'created': result['created'],
+			'compiled': result['compiled'],
 			'template': result['template'],
 			'url': get_tipp_url(id),
-			'qrurl': get_qr_url(id)
+			'qrurl': get_qr_url(id),
+			'content': ''
 		}
 
 		raw_path = Path(current_app.config['RAWPATH']) / f'{id}.md'
@@ -75,9 +83,9 @@ def tipp_details(id):
 
 		return tipp
 	else:
-		return ({'error': 404, 'msg': f'no tipp with id {id} found'}, 404)
+		return ({'error': 404, 'msg': f'unknown id'}, 404)
 
-@v1.route('/tipps/<string:id>', method=['PATCH'])
+@v1.route('/tipps/<string:id>', methods=['PATCH'])
 def tipp_compile(id):
 	template = request.args.get('template', default=None, type=str)
 
@@ -86,55 +94,75 @@ def tipp_compile(id):
 	if result:
 		if not template:
 			template = result['template']
-		compile_tipp(id, template=template)
+		timestamp = compile_tipp(id, template=template)
+		db.execute('UPDATE tipp SET compiled = ? WHERE id = ?', (timestamp, id,))
+		db.commit()
 		return {
 			'id': id,
-			'url': get_tipp_url(id),
-			'qrurl': get_qr_url(id)
+			'compiled': timestamp
 		}
 	else:
-		return ({'code': 400, 'msg': 'Unknown id!'}, 400)
-
-@v1.route('/tipps', methods=['POST'])
-def tipp_create():
-	if not authenticate():
-		return ({'code': 401, 'msg': 'Authentication failed!'}, 401)
-
-	template = request.args['template'] if 'template' in request.args else 'default'
-
-	body = request.data.decode().strip()
-	if len(body) == 0:
-		return ({'code': 400, 'msg': 'Body must not be empty!'}, 400)
-
-	id = create_tipp(body, template=template)
-
-	db = get_db()
-	db.execute('INSERT INTO tipp (id,user_id,template) VALUES (?, ?, ?)', (id, g.auth['user_id'], template))
-	db.commit()
-
-	return {
-		'id': id,
-		'url': get_tipp_url(id),
-		'qrurl': get_qr_url(id),
-	}
+		return ({'code': 400, 'msg': 'unknown id'}, 400)
 
 @v1.route('/tipps/<string:id>', methods=['DELETE'])
 def tipp_delete(id):
 	if not authenticate():
-		return ({'code': 401, 'msg': 'Authentication failed!'}, 401)
+		return ({'code': 401, 'msg': 'authentication failed'}, 401)
 	db = get_db()
-	db.execute('DELETE FROM tipp WHERE id = ?', (id,))
+	result = db.execute('SELECT user_id FROM tipp WHERE id = ?', (id,)).fetchone()
+	if not result:
+		return({'code': 404, 'msg': f'unknown id'}, 404)
+	if result['user_id'] != g.auth['user_id']:
+		return({'code': 403, 'msg': f'tipp owned by other token'}, 403)
+	db.execute('DELETE FROM tipp WHERE id = ? AND user_id = ?', (id, g.auth['user_id'],))
 	db.commit()
 	return ('', 204)
+
+@v1.route('/tipps', methods=['POST'])
+def tipp_create():
+	if not authenticate():
+		return ({'code': 401, 'msg': 'authentication failed'}, 401)
+
+	#template = request.args.get('template', default='default', type=str)
+
+	#content = request.data.decode().strip()
+	#if len(content) == 0:
+	#	content = request.form.get('content', default='').strip()
+	#if len(content) == 0:
+	#	return ({'code': 400, 'msg': 'body may not be empty'}, 400)
+
+	if 'content' not in request.json:
+		return ({'code': 400, 'msg': 'content may not be empty'}, 400)
+	content = str(request.json['content'])
+
+	template = 'default'
+	if 'template' in request.json:
+		template = str(request.json['template'])
+
+	id = create_tipp(content, template=template)
+
+	db = get_db()
+	db.execute('INSERT INTO tipp (id,user_id,template) VALUES (?, ?, ?)', (id, g.auth['user_id'], template,))
+	db.commit()
+
+	return ({
+		'id': id,
+		'created': datetime.now(),
+		'compiled': datetime.now(),
+		'template': template,
+		'url': get_tipp_url(id),
+		'qrurl': get_qr_url(id),
+		'content': content
+	}, 201)
 
 @v1.route('/tokens', methods=['POST'])
 def token_create():
 	if not authenticate():
-		return ({'code': 401, 'msg': 'Authentication failed!'}, 401)
+		return ({'code': 401, 'msg': 'authentication failed'}, 401)
 
 	data = request.json
 	if 'name' not in data or len(data['name']) == 0:
-		return ({'code': 400, 'msg': 'Provide a name for the new token: {"name": "..."}'}, 400)
+		return ({'code': 400, 'msg': 'missing token name'}, 400)
 	else:
 		token = generate_token()
 		user_id = g.auth['user_id']
@@ -143,7 +171,7 @@ def token_create():
 		db = get_db()
 		result = db.execute('SELECT id FROM user WHERE name = ?', (name,)).fetchone()
 		if result:
-			return ({'code': 400, 'msg': 'Tokenname already in use!'}, 400)
+			return ({'code': 409, 'msg': 'token name already in use'}, 409)
 		else:
 			db.execute('INSERT INTO user (name,token,ip,created_by) VALUES (?, ?, ?, ?)', (name, token, '0.0.0.0', user_id,))
 			db.commit()
